% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plpModelActiveCovariatesExtractor.R
\name{getLRModelCovariates}
\alias{getLRModelCovariates}
\title{get active model covariates of a logistic regression model}
\usage{
getLRModelCovariates(plpResults)
}
\arguments{
\item{plpResults}{a plp result object obtain by runPlp}
}
\value{
list of covariate names
}
\description{
get model covariate of a logistic model whose coefficients are different than 0
}
