% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformPlpData.R
\name{makeFriendlyNames}
\alias{makeFriendlyNames}
\title{Make friendly names}
\usage{
makeFriendlyNames(columnNames)
}
\arguments{
\item{columnNames}{column names}
}
\value{
new column names
}
\description{
remove punctuation marks and spaces
}
