% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformPlpData.R
\name{summarizeResults}
\alias{summarizeResults}
\title{summarize PLP results (Temporary function)}
\usage{
summarizeResults(s, evaluation)
}
\arguments{
\item{s}{a data frame with 'metric' and 'value' columns}

\item{evaluation}{a string}
}
\description{
Estimate external performance using external statistics and an internal dataset.
}
