% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformPlpData.R
\name{transformPlpDataToDataFrame}
\alias{transformPlpDataToDataFrame}
\title{Transform plp data to a data frame}
\usage{
transformPlpDataToDataFrame(plpData, populationSettings, outcomeId)
}
\arguments{
\item{plpData}{plpData}

\item{populationSettings}{population settings}

\item{outcomeId}{outcome id}
}
\value{
a data frame whose row ids or the original rowId of the input. The columns include features and the outcome
(in column 'outcome')
}
\description{
Transform plp data to a dense data frame. Set row names to the original rowId of the input
}
